Wrap _all_ async middleware stmts in try/catch
In Express v4.x, async middleware functions are required to explicitly
catch thrown Errors and call next(err) themselves.

https://expressjs.com/en/guide/error-handling.html

Wrapping the entire function body in a try/catch block is a simple way
to ensure this happens. Uncaught Errors crash the server completely: it
doesn't restart on its own. Doing this makes it less clear which
statements you expect to occassionally throw an error (in this case, the
database-querying ones), but uncaught Errors have an extreme cost here.